Make SQLite runtime driver-agnostic
