--- packages/bun-sqlite-runtime/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@yieldstar/bun-sqlite-runtime",
-  "version": "0.4.6",
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "bun build src/index.ts --outfile dist/index.js --target bun --packages=external"
-  },
-  "dependencies": {
-    "@yieldstar/core": "workspace:*",
-    "pino": "^9.9.0"
-  },
-  "devDependencies": {
-    "@types/bun": "^1.2.21"
-  }
-}
