+++ packages/sqlite-runtime/src/dao/task-queue-dao.ts
@@ -1,5 +1,5 @@
-import { Database } from "bun:sqlite";
 import type { WorkflowEvent } from "@yieldstar/core";
+import type { SqliteDriver } from "../sqlite-driver";
 
 class TaskRow {
   task_id!: number;
@@ -14,14 +14,14 @@ class CountRow {
 }
 
 export class TaskQueueDao {
-  private db: Database;
+  private db: SqliteDriver;
 
-  constructor(db: Database) {
+  constructor(db: SqliteDriver) {
     this.db = db;
   }
 
   setupDb() {
-    this.db.exec(`
+    this.db.run(`
       CREATE TABLE IF NOT EXISTS task_queue (
         task_id INTEGER PRIMARY KEY AUTOINCREMENT,
         workflow_id TEXT NOT NULL,
@@ -34,7 +34,7 @@ export class TaskQueueDao {
   }
 
   insertTask(event: WorkflowEvent) {
-    const query = this.db.prepare(
+    const query = this.db.query(
       `INSERT INTO task_queue (workflow_id, execution_id, params, context)
       VALUES ($workflowId, $executionId, $params, $context)`
     );
@@ -50,31 +50,30 @@ export class TaskQueueDao {
   }
 
   getNextTask() {
-    const query = this.db.prepare(
+    const query = this.db.query<TaskRow>(
       `SELECT * FROM task_queue WHERE visible_from < strftime('%s', 'now') ORDER BY task_id LIMIT 1`
     );
-    return query.get() as TaskRow | undefined;
+    return query.get();
   }
 
   updateTaskVisibility(taskId: number, visibleFrom: number) {
-    const query = this.db.prepare(
+    const query = this.db.query(
       "UPDATE task_queue SET visible_from = $visibleFrom WHERE task_id = $taskId"
     );
     query.run({ $taskId: taskId, $visibleFrom: visibleFrom });
   }
 
   deleteTaskById(id: number) {
-    const query = this.db.prepare(
+    const query = this.db.query(
       `DELETE FROM task_queue WHERE task_id = $taskId`
     );
     query.run({ $taskId: id });
   }
 
   getTaskCount() {
-    const query = this.db.prepare(
+    const query = this.db.query<CountRow>(
       `SELECT COUNT(*) as count FROM task_queue WHERE visible_from < strftime('%s', 'now')`
     );
-    const result = query.get() as CountRow;
-    return result.count;
+    return query.get()!.count;
   }
 }
