+++ packages/sqlite-runtime/src/index.ts
@@ -4,4 +4,11 @@ export { SqliteSchedulerClient } from "./sqlite-scheduler";
 export { SqliteEventLoop } from "./sqlite-event-loop";
 export { SqliteTaskQueueClient } from "./sqlite-task-queue";
 export { SqliteTimersClient } from "./sqlite-timers";
-export { createSqliteDb } from "./sqlite-db";
+export type {
+  SqliteDriver,
+  SqliteParameters,
+  SqliteParameterValue,
+  SqliteRunResult,
+  SqliteStatement,
+} from "./sqlite-driver";
+export type { SqliteDbOpts } from "./sqlite-db";
