+++ tsconfig.json
@@ -8,7 +8,7 @@
     { "path": "./packages/test-runtime" },
     { "path": "./packages/test-utils" },
     { "path": "./packages/bun-http-server" },
-    { "path": "./packages/bun-sqlite-runtime" },
+    { "path": "./packages/sqlite-runtime" },
     { "path": "./packages/bun-worker-invoker" }
   ]
 }
